fix: PR #74 回归修复 — 算术溢出/作用域/分配/退款通知等 13 项

+++ fengyu-admin/src/actions/cards.ts
@@ -1002,7 +1002,7 @@ export const createRechargeOrder = withPermission(
       saleOrderId = await db.transaction(async (tx) => {
         const idRows = await tx.execute(sql`
           WITH lock AS (
-            SELECT pg_advisory_xact_lock(hashtext('sale_order_id_gen'))
+            SELECT pg_advisory_xact_lock(hashtext('sale_order_id_gen')::bigint)
           )
           SELECT 'FY-XSD-WX-' || to_char(NOW(), 'YYMMDD') ||
             LPAD(
